# [WIP] Attempting to speedup tests by pickling cpython output