… tests/cache/pyout_cache-3.4.3.pickle