This is an initial attempt to try to speed up the tests, by pickling the Python output (which normally doesn't change).

It's enabled by an environment variable USE_PYOUT_CACHE, if this is present it will try to load the cache from a pickle file, and will update the cache file at the end of each test class.

Not sure if this is worth yet because of the hassle of maintaining the cache files, I'll do some more tests.

Running for `tests.datatypes.test_list.ListTests`, the time went from `8s` to `5s` (~40% speedup, encouraging!).
Running for `tests.datatypes.test_list`, the time went from `205s` to `165s` (~20% speedup, not much).

I suspect the latter case gave a lesser speedup because it's updating the whole cache pickle file each time it runs a test case, so I'd like to try next splitting the pickle file per test classes.