@eliasdorneles 2016-11-13
Of course, to get the benefits of this in Travis, we'll have to run locally in a similar environment first, at least for most of the tests.

We'd have to find a decent way of keeping these cache files updated.

I'll appreciate any feedback. :)