@eliasdorneles 2016-11-15
@freakboy3742 looks reenabling the verbose output still doesn't fix Travis builds timing out. =/

I haven't dug much, but we might try just running separate jobs to test each subset of the tests: https://docs.travis-ci.com/user/speeding-up-the-build/
If I understood correctly, it's okay to have several jobs, as long as they don't take too long.