@cflee 2016-11-16
@eliasdorneles going back to verbose output doesn't help due to the overall 50 min limit:

> Travis CI has specific time limits for each job, and will stop the build and and add an error message to the build log in the following situations:

> * A job takes longer than 50 minutes on travis-ci.org

([source](https://docs.travis-ci.com/user/customizing-the-build#Build-Timeouts))

Root cause: this policy seems to be in place as far back as 2014 (https://github.com/travis-ci/travis-ci/issues/2736#issue-41559910) – but rust-lang discovered in Jul 2015 (https://github.com/travis-ci/travis-ci/issues/4521) that the 50 min timeout was not working properly on the container infrastructure which we are on. That explains the 120 min limit we saw previously; I guess Travis recently patched that bug.

I think you're right, splitting into several jobs each under 50 min is the workaround.

/cc @freakboy3742 in case you are looking to invent more new swear words 🙃