@eliasdorneles 2016-11-17
I'm going to close this because I don't think the hassle of maintaining the pickle files is worth the little more performance it gives.