@freakboy3742 2016-11-18
If I get a moment over the weekend, I'm going to try an experiment that is related to this. Instead of pickling, just use the disk cache: Roll out tests to the file system and check them in, along with the pyc files. 

If you don't see any feedback after this weekend ,feel free to try this experiment yourself :-)